+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- lookup button form element
+
 ### Changed
 
 - Section collapsed/expanded state across persists across autosave/drafts